reenable test for existing policy

+++ scripts/moonraker-ensure-policykit-rules.sh
@@ -6,21 +6,21 @@ POLKIT_USR_DIR="/usr/share/polkit-1/rules.d"
 MOONRAKER_UNIT="/etc/systemd/system/moonraker.service"
 
 ensure_moonraker_policiykit_rules() {
-	# if [[ -e ${POLKIT_USR_DIR}/moonraker.rules ]]
-	# then
-	# 	echo -e "\n\n###### Moonraker policy exists, skipping policykit script."
-	# 	return
-	# fi
-	# if [[ -e ${POLKIT_DIR}/moonraker.rules ]]
-	# then
-	# 	echo -e "\n\n###### Moonraker policy exists, skipping policykit script."
-	# 	return
-	# fi
-	# if [[ -e ${POLKIT_LEGACY_DIR}/10-moonraker.pkla ]]
-	# then
-	# 	echo -e "\n\n###### Moonraker legacy policy exists, skipping policykit script."
-	# 	return
-	# fi
+	if [[ -e ${POLKIT_USR_DIR}/moonraker.rules ]]
+	then
+		echo -e "\n\n###### Moonraker policy exists, skipping policykit script."
+		return
+	fi
+	if [[ -e ${POLKIT_DIR}/moonraker.rules ]]
+	then
+		echo -e "\n\n###### Moonraker policy exists, skipping policykit script."
+		return
+	fi
+	if [[ -e ${POLKIT_LEGACY_DIR}/10-moonraker.pkla ]]
+	then
+		echo -e "\n\n###### Moonraker legacy policy exists, skipping policykit script."
+		return
+	fi
 	if [[ -e /home/pi/moonraker/scripts/set-policykit-rules.sh ]]
 	then
 		cp /home/pi/moonraker/scripts/set-policykit-rules.sh /tmp/set-policykit-rules.sh
